+++ app/routes/index.js
@@ -4,17 +4,6 @@ export default Ember.Route.extend({
 
   kredits: Ember.inject.service(),
 
-  beforeModel(transition) {
-    const kredits = this.get('kredits');
-    return kredits.initEthProvider().then(() => {
-      if (kredits.get('accountNeedsUnlock')) {
-        if (confirm('It looks like you have an Ethereum wallet available. Please unlock your account.')) {
-          transition.retry();
-        }
-      }
-    });
-  },
-
   model() {
     let newContributor = Ember.getOwner(this).lookup('model:contributor');
     newContributor.set('kind', 'person');
